requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
cx-Freeze==6.15.11
cx_Logging==3.2.0
darkdetect==0.8.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
lief==0.14.1
Markdown==3.6
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
tkhtmlview==0.3.0
tzdata==2024.1
urllib3==2.2.1
